+++ src/main/java/org/erachain/gui/models/WalletTransactionsTableModel.java
@@ -269,6 +269,9 @@ public class WalletTransactionsTableModel extends TableModelCls<Tuple2<String, S
 
     @SuppressWarnings("unchecked")
     public synchronized void syncUpdate(Observable o, Object arg) {
+        if (Controller.getInstance().wallet.database == null)
+            return;
+
         ObserverMessage message = (ObserverMessage) arg;
 
         //CHECK IF NEW LIST
@@ -339,6 +342,12 @@ public class WalletTransactionsTableModel extends TableModelCls<Tuple2<String, S
         } else if (message.getType() == ObserverMessage.ADD_UNC_TRANSACTION_TYPE) {
             // INCOME
 
+            long period = NTP.getTime() - this.timeUpdate;
+            if (period < 2000) //Gui.PERIOD_UPDATE)
+                return;
+            this.timeUpdate = NTP.getTime();
+            needUpdate = false;
+
             this.fireTableDataChanged();
             if (true)
                 return;
@@ -407,6 +416,9 @@ public class WalletTransactionsTableModel extends TableModelCls<Tuple2<String, S
 
     public void addObservers() {
 
+        if (!Controller.getInstance().doesWalletDatabaseExists())
+            return;
+
         //REGISTER ON WALLET TRANSACTIONS
         Controller.getInstance().getWallet().database.getTransactionMap().addObserver(this);
         // for UNCONFIRMEDs
@@ -418,6 +430,9 @@ public class WalletTransactionsTableModel extends TableModelCls<Tuple2<String, S
 
 
     public void removeObservers() {
+
+        if (Controller.getInstance().doesWalletDatabaseExists())
+            return;
 
         Controller.getInstance().getWallet().database.getTransactionMap().deleteObserver(this);
         DCSet.getInstance().getTransactionMap().addObserver(this);
